+++ netmedic/__init__.py
@@ -7,7 +7,8 @@ Author : Birditch
 License: MIT
 """
 
-__version__ = "1.0.0"
+__version__ = "1.0.1b1"
+__display_version__ = "1.0.1 preview beta1"
 __minimum_python__ = "3.10"
 __author__ = "Birditch"
 __license__ = "MIT"